refactor(cli): extract doctor/network/safety/machine/history to *_cli.py (#495)
Phase 1 (Wave 2) of #495. Moves the diagnostics + remote-machine domains out
of the __main__ monolith into four modules, mirroring the diff_cli/prompts_cli
pattern:

- doctor_cli.py  — doctor (+ render helpers, _confirm) and network status
- safety_cli.py  — safety check/status/notify-unattended-block/logs/install/tooldefs
- machine_cli.py — machine add/remove/list
- history_cli.py — history list/show/resume

Pure relocation; CLI surface (commands, flags, help, behavior) unchanged.
doctor reads the hooks-domain drift helpers (get_hooks_source,
_managed_hook_files, _managed_file_state, skill_drift, CLAUDE_SKILLS_DIR) via
a function-local deferred import — they stay in __main__ with the still-unextracted
hooks domain, so this preserves single-source-of-truth rather than duplicating.
Tests that patched moved symbols repointed to the new modules.

Built by dotdev.dev
